requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
discord.py==2.0.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
yarl==1.8.1
